% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_balance}
\alias{compute_balance}
\title{Compute the imbalance for a given design matrix}
\usage{
compute_balance(D, weights, sample_counts, target_counts)
}
\arguments{
\item{D}{Design matrix, output from \code{\link{create_design_matrix}}}

\item{weights}{Vector of weights}

\item{sample_counts}{Vector of sample counts for each cell}

\item{target_counts}{Vector of target counts for each cell}
}
\description{
Compute the imbalance for a given design matrix
}
\keyword{internal}
