% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multical.R
\name{create_cells}
\alias{create_cells}
\title{Create data frame with all distinct cells, and their sample and target counts}
\usage{
create_cells(formula, target_count, data)
}
\arguments{
\item{formula}{Right-hand-side formula of the form \code{~ covariates}
specifying the covariates to calibrate on}

\item{target_count}{Optional. Name of column in \code{pop_data} giving the
count or weight for each row. Defaults to 1 per row (row-count semantics).}
}
\description{
Create data frame with all distinct cells, and their sample and target counts
}
\keyword{internal}
