% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multical.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Creates the underlying design matrix to solve the multilevel calibration
optimization problem}
\usage{
create_design_matrix(cells, order)
}
\arguments{
\item{cells}{Dataframe of distinct cells}

\item{order}{Integer. What order interactions to balance}
}
\description{
Creates the underlying design matrix to solve the multilevel calibration
optimization problem
}
\keyword{internal}
