% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{create_rake_Pmat}
\alias{create_rake_Pmat}
\title{Create matrix for quadratic term in QP}
\usage{
create_rake_Pmat(D, sample_counts, lambda)
}
\arguments{
\item{D}{Design matrix, output from \code{\link{create_design_matrix}}}

\item{sample_counts}{Vector of sample counts for each cell}

\item{lambda}{Numeric. Hyperparameter}
}
\description{
Create matrix for quadratic term in QP
}
\keyword{internal}
