% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cell}
\alias{data_cell}
\title{Fake data at the cell level}
\format{
A data frame with 91 rows and 6 variables:
\describe{
\item{X1, X2, X3, X4}{Categorical variables}
\item{sample_count}{Number of respondents in the cell}
\item{target_count}{Total population size in the cell}
\item{y}{Total answers of 1 to a binary survey question in the cell}
}
}
\usage{
data_cell
}
\description{
Fake data at the cell level
}
\keyword{datasets}
