% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_individual}
\alias{data_individual}
\title{Fake data at the individual level}
\format{
A data frame with 2500 rows and 6 variables:
\describe{
\item{X1, X2, X3, X4}{Categorical variables}
\item{response}{Whether the individual responded to the survey}
\item{y}{Binary answer to survey question}
}
}
\usage{
data_individual
}
\description{
Fake data at the individual level
}
\keyword{datasets}
