% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_balance}
\alias{get_balance}
\title{Get the difference between the reweighted sample and the target}
\usage{
get_balance(x, order, ...)
}
\arguments{
\item{x}{A \code{\link{multical}} object}

\item{order}{Integer. Order of interactions to include in the balance measure}

\item{...}{Additional arguments (unused)}
}
\value{
A data frame with the difference for each term and value of the
hyper-parameter
}
\description{
For each interaction term, computes the difference between the reweighted
sample count and the target count.
}
