% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_balance_v_sample_size}
\alias{get_balance_v_sample_size}
\title{Compute the overall level of balance versus the effective sample size}
\usage{
get_balance_v_sample_size(output, order)
}
\arguments{
\item{output}{Output of \code{\link{multical}}}

\item{order}{Integer. Order of interactions to include in the balance measure}
}
\value{
A data frame with the imbalance and effective sample size for each
value of the hyper-parameter
}
\description{
For each value of the hyper-parameter, computes the effective sample size
and the overall root sum of squared differences in all interactions,
normalized by the size of the target population
}
