% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.multical}
\alias{plot.multical}
\title{Plot balance vs. effective sample size for a multical object}
\usage{
\method{plot}{multical}(x, order = NULL, ...)
}
\arguments{
\item{x}{A \code{multical} object}

\item{order}{Integer. Order of interactions for the balance measure;
defaults to \code{x$order}}

\item{...}{Passed to \code{\link[graphics]{plot}} when \pkg{ggplot2} is
unavailable}
}
\description{
Calls \code{\link{get_balance_v_sample_size}} and plots the trade-off curve,
highlighting the auto-selected default lambda in red.
}
