% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.multical}
\alias{print.multical}
\title{Print a multical object}
\usage{
\method{print}{multical}(x, ...)
}
\arguments{
\item{x}{A \code{multical} object}

\item{...}{Ignored}
}
\description{
Print a multical object
}
