% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{update_rake_Pmat}
\alias{update_rake_Pmat}
\title{Update the quadratic term matrix with a new hyperparameter value}
\usage{
update_rake_Pmat(P, sample_counts, lambda)
}
\arguments{
\item{P}{Original quadratic term matrix}

\item{sample_counts}{Vector of sample counts for each cell}

\item{lambda}{Numeric. Hyperparameter}
}
\description{
Update the quadratic term matrix with a new hyperparameter value
}
\keyword{internal}
