% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multical.R
\name{calibrate_}
\alias{calibrate_}
\title{Internal function to perform multilevel calibration}
\usage{
calibrate_(
  cells,
  sample_counts,
  target_counts,
  order = NULL,
  lambda = 1,
  lambda_max = NULL,
  n_lambda = 100,
  lambda_min_ratio = 1e-05,
  lowlim = 0,
  uplim = Inf,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cells}{Dataframe of distinct cells}

\item{sample_counts}{Vector of sample counts for each cell}

\item{target_counts}{Vector of target counts for each cell}

\item{order}{Integer. What order interactions to balance. Default is all orders}

\item{lambda}{Numeric. Regularization hyperparamter, by default fits weights
for a range of values}

\item{lambda_max}{Numeric. Maximum hyperparameter to fit weights with,
default is the root sum of squared differences between the (unweighted) sample and the target}

\item{n_lambda}{Integer. Number of hyper-parameters to fit weights for, from
lambda_max to lambda_max * lambda_min_ratio, equally spaced on the log scale. Default, 20}

\item{lambda_min_ratio}{Numeric. Ratio of min to max lambda to consider.}

\item{lowlim}{Lower bound on weights, default 0}

\item{uplim}{Upper bound on weights, default Inf}

\item{verbose}{Boolean. Show optimization information, default False}

\item{...}{Additional parameters for osqp}
}
\description{
Internal function to perform multilevel calibration
}
\keyword{internal}
