% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{create_cells}
\alias{create_cells}
\title{Create DataFrame with all distinct cells, and their sample and target counts}
\usage{
create_cells(formula, target_count, data)
}
\arguments{
\item{formula}{Formula of the form \code{sample_count ~ covariates}, where
\code{sample_count} is whether or not an individual responded
(with individual-level data) or the number of respondents in the cell
(with cell-level data) and \code{covariates} are the covariates to calibrate
on}

\item{target_count}{Name of column with indicators for whether an individual
is in the target population (with individual-level data) or the target counts for each cell (with cell-level data)}

\item{data}{Dataframe with covariate information, sample and target counts}
}
\description{
Create DataFrame with all distinct cells, and their sample and target counts
}
\keyword{internal}
