% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{create_rake_constraints}
\alias{create_rake_constraints}
\title{Create underlying constraints for optimization problem}
\usage{
create_rake_constraints(
  cells,
  D,
  sample_counts,
  target_counts,
  lowlim,
  uplim,
  verbose
)
}
\arguments{
\item{cells}{Dataframe of distinct cells}

\item{D}{Design matrix, output from \code{\link{create_design_matrix}}}

\item{sample_counts}{Vector of sample counts for each cell}

\item{target_counts}{Vector of target counts for each cell}

\item{lowlim}{Lower bound on weights, default 0}

\item{uplim}{Upper bound on weights, default Inf}

\item{verbose}{Boolean. Show optimization information, default False}
}
\description{
Create underlying constraints for optimization problem
}
\keyword{internal}
