% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_balance}
\alias{get_balance}
\title{Get the difference between the reweighted sample and the target}
\usage{
get_balance(output, order)
}
\arguments{
\item{output}{Output of \code{\link{multical}}}

\item{order}{Integer. Order of interactions to include in the balance measure}
}
\value{
A data frame with the difference for each term and value of the
hyper-parameter
}
\description{
For each interaction term, computes the difference between the reweighted
sample count and the target count
}
