% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multical.R
\name{multical}
\alias{multical}
\title{Calibrate sample to target via multilevel calibration}
\usage{
multical(
  formula,
  target_count,
  data,
  order = NULL,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 20,
  lambda_min_ratio = 1e-05,
  lowlim = 0,
  uplim = Inf,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{Formula of the form \code{sample_count ~ covariates}, where
\code{sample_count} is whether or not an individual responded
(with individual-level data) or the number of respondents in the cell
(with cell-level data) and \code{covariates} are the covariates to calibrate
on}

\item{target_count}{Name of column with indicators for whether an individual
is in the target population (with individual-level data) or the target counts for each cell (with cell-level data)}

\item{data}{Dataframe with covariate information, sample and target counts}

\item{order}{Integer. What order interactions to balance. Default is all orders}

\item{lambda}{Numeric. Regularization hyperparamter, by default fits weights
for a range of values}

\item{lambda_max}{Numeric. Maximum hyperparameter to fit weights with,
default is the root sum of squared differences between the (unweighted) sample and the target}

\item{n_lambda}{Integer. Number of hyper-parameters to fit weights for, from
lambda_max to lambda_max * lambda_min_ratio, equally spaced on the log scale. Default, 20}

\item{lambda_min_ratio}{Numeric. Ratio of min to max lambda to consider.}

\item{lowlim}{Lower bound on weights, default 0}

\item{uplim}{Upper bound on weights, default Inf}

\item{verbose}{Boolean. Show optimization information, default False}

\item{...}{Additional parameters for osqp}
}
\value{
data frame with the weight for each distinct cell, for each value of
the hyperparameter \code{lambda}. Note: the output data frame may have the
cells in a different order than in \code{data}. Be sure to join the output
with \code{data} on the variables to map the weights to the data accurately.
}
\description{
Finds weights that exactly calibrate first order margins between respondents and the target population. Requires individual-level of cell-level data.
}
